feat(superchain): switch to JDK 20 (#4082)
Switch to a more recent version of the JDK. `jsii-pacmak` will keep on targeting Java 1.8, so this doesn't affect compatibility of the generated class files.

What this will gain us is access to a more recent, optimized `javadoc` tool, which completes a lot faster than the one from JDK 8.



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0