chore: download Corretto key at build time
The Corretto key recently expired, bringing down a whole host of
projects that are relying on Superchain. In the mean time, Corretto team
extended the lifetime of this key and are publishing it at a fixed URL.

In this PR, download the key at build time instead of hardcoding it
into the repository.

Yes, this is not as secure, as the key download could be intercepted.
But this risk is probably low, while the availability risk of expired
keys is real and present.